move 'add run method' code into it's own componenet
